fixed same data calculation for reminder emails
