+++ InterneTaakAfhandeling.Web.Server/Features/InternetakenOverviews/Shared/Urgentie/UrgentieBerekenService.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.Options;
+﻿using InterneTaakAfhandeling.Common.Helpers;
+using Microsoft.Extensions.Options;
 
 namespace InterneTaakAfhandeling.Web.Server.Features.InternetakenOverviews.Shared.Urgentie;
 
@@ -17,9 +18,9 @@ public class UrgentieBerekenService(
         }
 
         var afhandeltermijn = TimeSpan.FromHours(_options.AfhandeltermijnUren);
-        var streefdatum = AddWeekdayHours(contactDatum.Value, afhandeltermijn);
+        var streefdatum = WerkdagenCalculator.AddWeekdayHours(contactDatum.Value, afhandeltermijn);
         var now = DateTimeOffset.UtcNow.ToOffset(streefdatum.Offset);
-        var resterendeWerkdagUren = BerekenResterendeWerkdagUren(now, streefdatum);
+        var resterendeWerkdagUren = WerkdagenCalculator.BerekenResterendeWerkdagUren(now, streefdatum);
 
         var status = resterendeWerkdagUren switch
         {
@@ -56,88 +57,4 @@ public class UrgentieBerekenService(
         };
     }
 
-    /// <summary>
-    /// Adds the given number of hours to a start time, skipping weekends entirely (24/7 on weekdays).
-    /// </summary>
-    internal static DateTimeOffset AddWeekdayHours(DateTimeOffset start, TimeSpan hoursToAdd)
-    {
-        var current = start;
-        var remainingHours = hoursToAdd.TotalHours;
-
-        while (remainingHours > 0)
-        {
-            if (IsWeekend(current))
-            {
-                current = SkipToNextWeekday(current);
-                continue;
-            }
-
-            var endOfDay = new DateTimeOffset(current.Date.AddDays(1), current.Offset);
-            var hoursLeftInDay = (endOfDay - current).TotalHours;
-
-            if (remainingHours <= hoursLeftInDay)
-            {
-                current = current.AddHours(remainingHours);
-                remainingHours = 0;
-            }
-            else
-            {
-                remainingHours -= hoursLeftInDay;
-                current = endOfDay;
-            }
-        }
-
-        return current;
-    }
-
-    /// <summary>
-    /// Calculates the number of weekday hours (excluding Saturday/Sunday) between now and the streefdatum.
-    /// Returns negative values when the streefdatum has passed.
-    /// </summary>
-    internal static double BerekenResterendeWerkdagUren(DateTimeOffset now, DateTimeOffset streefdatum)
-    {
-        if (now >= streefdatum)
-        {
-            return -BerekenWerkdagUrenTussen(streefdatum, now);
-        }
-
-        return BerekenWerkdagUrenTussen(now, streefdatum);
-    }
-
-    private static double BerekenWerkdagUrenTussen(DateTimeOffset start, DateTimeOffset end)
-    {
-        var current = start;
-        var totalHours = 0.0;
-
-        while (current < end)
-        {
-            if (IsWeekend(current))
-            {
-                current = SkipToNextWeekday(current);
-                continue;
-            }
-
-            var endOfDay = new DateTimeOffset(current.Date.AddDays(1), current.Offset);
-            var dayEnd = endOfDay < end ? endOfDay : end;
-
-            totalHours += (dayEnd - current).TotalHours;
-            current = endOfDay;
-        }
-
-        return totalHours;
-    }
-
-    private static bool IsWeekend(DateTimeOffset date) =>
-        date.DayOfWeek is DayOfWeek.Saturday or DayOfWeek.Sunday;
-
-    private static DateTimeOffset SkipToNextWeekday(DateTimeOffset date)
-    {
-        var next = date.Date.AddDays(1);
-        while (next.DayOfWeek is DayOfWeek.Saturday or DayOfWeek.Sunday)
-        {
-            next = next.AddDays(1);
-        }
-
-        return new DateTimeOffset(next, date.Offset);
-    }
 }
