+++ InterneTaakAfhandeling.Web.Server/Features/InternetakenOverviews/Shared/Urgentie/UrgentieOptions.cs
@@ -1,4 +1,4 @@
-﻿namespace InterneTaakAfhandeling.Web.Server.Features.InternetakenOverviews.Shared.Urgentie;
+namespace InterneTaakAfhandeling.Web.Server.Features.InternetakenOverviews.Shared.Urgentie;
 
 public class UrgentieOptions
 {
